common openapi gen
