+++ pkg/apis/dashboardsnapshot/v0alpha1/zz_generated.openapi_violation_exceptions.list
